+++ R/ternary_hull.R
@@ -8,8 +8,8 @@ NULL
 setMethod(
   f = "ternary_hull",
   signature = c(x = "numeric", y = "numeric", z = "numeric"),
-  definition = function(x, y, z, ...) {
-    coords <- coordinates_ternary(x, y, z)
+  definition = function(x, y, z, center = FALSE, scale = FALSE, ...) {
+    coords <- coordinates_ternary(x, y, z, center = center, scale = scale)
     hull <- grDevices::chull(coords)
     graphics::polygon(x = coords$x[hull], y = coords$y[hull], ...)
 
@@ -24,9 +24,10 @@ setMethod(
 setMethod(
   f = "ternary_hull",
   signature = c(x = "ANY", y = "missing", z = "missing"),
-  definition = function(x, ...) {
+  definition = function(x, center = FALSE, scale = FALSE, ...) {
     xyz <- grDevices::xyz.coords(x)
-    coords <- methods::callGeneric(x = xyz$x, y = xyz$y, z = xyz$z, ...)
+    coords <- methods::callGeneric(x = xyz$x, y = xyz$y, z = xyz$z,
+                                   center = center, scale = scale, ...)
     invisible(coords)
   }
 )