+++ R/ternary_text.R
@@ -8,8 +8,9 @@ NULL
 setMethod(
   f = "ternary_text",
   signature = c(x = "numeric", y = "numeric", z = "numeric"),
-  definition = function(x, y, z, labels = seq_along(x), ...) {
-    coords <- coordinates_ternary(x, y, z)
+  definition = function(x, y, z, center = FALSE, scale = FALSE,
+                        labels = seq_along(x), ...) {
+    coords <- coordinates_ternary(x, y, z, center = center, scale = scale)
     graphics::text(x = coords, labels = labels, ...)
 
     invisible(list(x = x, y = y, z = z))
@@ -22,10 +23,12 @@ setMethod(
 setMethod(
   f = "ternary_text",
   signature = c(x = "ANY", y = "missing", z = "missing"),
-  definition = function(x, labels = seq_along(x$x), ...) {
+  definition = function(x, center = FALSE, scale = FALSE, labels = seq_along(x$x), ...) {
     x <- grDevices::xyz.coords(x)
     force(labels)
-    coords <- methods::callGeneric(x = x$x, y = x$y, z = x$z, labels = labels, ...)
+    coords <- methods::callGeneric(x = x$x, y = x$y, z = x$z,
+                                   center = center, scale = scale,
+                                   labels = labels, ...)
     invisible(coords)
   }
 )